pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "opencount-ci"
version = "0.2.0"
description = "Class-agnostic object counting with confidence intervals (no ML)."
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [{ name = "Oliger Shehi" }]

dependencies = [
  "opencv-python>=4.8",
  "numpy>=1.22,<1.25",
  "Pillow>=10.0",
]

[project.scripts]
opencount = "opencount_ci.cli:main"
